refactor(wordleverse): actions
