+++ jsconfig.json
@@ -1,7 +1,8 @@
 {
   "compilerOptions": {
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./src/*"],
+      "@wordleverse/*": ["./src/app/(pages)/games/wordleverse/*"]
     }
   }
 }
