--- src/app/(pages)/games/wordleverse/Status.js
@@ -1,8 +1,0 @@
-const Status = Object.freeze({
-  Absent: "absent",
-  Default: "default",
-  Present: "present",
-  Correct: "correct",
-});
-
-export default Status;
