+++ src/app/(pages)/games/wordleverse/_context/reducer/playAgain.js
@@ -1,6 +1,6 @@
 import { random, cloneDeep } from "lodash";
 import words from "@/data/words.json";
-import { defaultState } from "..";
+import { defaultState } from "@wordleverse/_lib/defaults";
 
 export const playAgain = ({ expert }) => ({
   ...cloneDeep(defaultState),
