+++ src/app/(pages)/games/wordleverse/history/_actions/getHistory.js
@@ -1,0 +1,51 @@
+"use server";
+
+import { getSession } from "@/lib/auth";
+import prisma from "@/lib/prisma";
+
+export const getHistory = async (options = {}) => {
+   const session = await getSession();
+    if (!session?.user) {
+      return { error: "Unauthorized", status: 401 };
+    }
+    const userId = session.user.id;
+  
+    try {
+      // Get user's game history with streak information
+      const games = await prisma.wordleHistory.findMany({
+        where: {
+          userId,
+        },
+        orderBy: {
+          date: "desc",
+        },
+      });
+    
+      // Get user's streak information
+      const user = await prisma.user.findUnique({
+        where: { id: userId },
+        select: { streak, maxStreak },
+      });
+
+      const history = {
+        games,
+        streak: user?.streak || 0,
+        maxStreak: user?.maxStreak || 0,
+      }
+
+      // If requested, include available dates
+      const includeAvailableDates = options.includeAvailableDates === true;
+      if (includeAvailableDates) {
+        const availableDates = await getAvailableDates(session.user.id);
+        return {
+          ...history,
+          availableDates
+        };
+      }
+      
+      return history;
+    } catch (error) {
+      console.error("Error getting history:", error);
+      return { error: "Failed to get history", status: 500 };
+    }
+}
