+++ src/app/(pages)/games/wordleverse/history/_components/styled.js
@@ -1,6 +1,7 @@
 "use client";
 
 import styled from "@emotion/styled";
+import Link from "next/link";
 
 export const Container = styled.div`
   display: flex;
